Update migration file to change the data type of the "saldo" column from unsigned integer to unsigned big integer

+++ database/migrations/2024_07_26_201455_kas.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string("tanggal");
             $table->string("jenis");
             $table->unsignedInteger("jumlah");
-            $table->unsignedInteger("saldo");
+            $table->unsignedBigInteger("saldo");
             $table->string("keterangan");
             $table->timestamps();
         });
